refactor(core): move queue item state ownership into the runtime layer
