+++ .changeset/queue-item-layering.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/core": patch
+---
+
+refactor: move queue item state ownership into the runtime layer.
